+++ src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/PigZombieShop.java
@@ -1,19 +1,26 @@
 package com.nisovin.shopkeepers.shopobjects.living.types;
 
+import java.util.List;
+
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
-import org.bukkit.entity.EntityType;
 import org.bukkit.entity.PigZombie;
+import org.bukkit.entity.Player;
+import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 
+import com.nisovin.shopkeepers.Settings;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.shopobjects.living.LivingShops;
 import com.nisovin.shopkeepers.shopobjects.living.SKLivingShopObject;
 import com.nisovin.shopkeepers.shopobjects.living.SKLivingShopObjectType;
+import com.nisovin.shopkeepers.ui.defaults.EditorHandler;
+import com.nisovin.shopkeepers.util.ItemUtils;
 
-public class PigZombieShop extends SKLivingShopObject {
+public class PigZombieShop extends SKLivingShopObject<PigZombie> {
 
+	// TODO use BabyableShop as base once there is a common interface for this inside bukkit
 	private boolean baby = false;
 
 	public PigZombieShop(	LivingShops livingShops, SKLivingShopObjectType<PigZombieShop> livingObjectType,
@@ -34,29 +41,54 @@ public class PigZombieShop extends SKLivingShopObject {
 	}
 
 	@Override
-	public PigZombie getEntity() {
-		assert super.getEntity().getType() == EntityType.PIG_ZOMBIE;
-		return (PigZombie) super.getEntity();
+	protected void onSpawn(PigZombie entity) {
+		super.onSpawn(entity);
+		this.applyBaby(entity);
 	}
 
-	// SUB TYPES
+	// EDITOR ACTIONS
 
 	@Override
-	protected void applySubType() {
-		super.applySubType();
-		if (!this.isActive()) return;
-		this.getEntity().setBaby(baby);
+	public List<EditorHandler.Button> getEditorButtons() {
+		List<EditorHandler.Button> editorButtons = super.getEditorButtons(); // assumes modifiable
+		editorButtons.add(new EditorHandler.ActionButton(shopkeeper) {
+			@Override
+			public ItemStack getIcon(EditorHandler.Session session) {
+				return getBabyEditorItem();
+			}
+
+			@Override
+			protected boolean runAction(InventoryClickEvent clickEvent, Player player) {
+				cycleBaby();
+				return true;
+			}
+		});
+		return editorButtons;
 	}
 
-	@Override
-	public ItemStack getSubTypeItem() {
-		return new ItemStack(Material.ZOMBIE_PIGMAN_SPAWN_EGG, 1);
-	}
+	// BABY STATE
 
-	@Override
-	public void cycleSubType() {
+	public void setBaby(boolean baby) {
+		this.baby = baby;
 		shopkeeper.markDirty();
-		baby = !baby;
-		this.applySubType();
+		this.applyBaby(this.getEntity()); // null if not active
+	}
+
+	protected void applyBaby(PigZombie entity) {
+		if (entity == null) return;
+		entity.setBaby(baby);
+	}
+
+	public void cycleBaby() {
+		this.setBaby(!baby);
+	}
+
+	protected ItemStack getBabyEditorItem() {
+		// TODO use mob-specific spawn egg (if available; some mobs (illusioner) don't have a spawn egg)?
+		// on the other hand: using a single item consistently for the editor icon has benefits as well
+		ItemStack iconItem = new ItemStack(Material.EGG);
+		// TODO use more specific text
+		ItemUtils.setItemStackNameAndLore(iconItem, Settings.msgButtonType, Settings.msgButtonTypeLore);
+		return iconItem;
 	}
 }